pyproject.toml
[tool.poetry]
name = "ominimo-api-gateway"
version = "0.1.0"
description = ""
authors = ["Mladen"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.120.0,<0.121.0"
uvicorn = {version = "0.24.0", extras = ["standard"]}
httpx = "0.25.2"
pyyaml = ">=6.0.3,<7.0.0"
python-multipart = "0.0.6"
pydantic = "^2.8"            # (no exact pin)
pydantic-settings = ">=2.11.0,<3.0.0"
python-dotenv = "^1.0"
prometheus-client = "^0.20"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
black = "25.9.0"
pylint = "4.0.2"
pycln = "2.6.0"
insurance-model-a = { git = "ssh://git@github.com/Mladja99/mock_models.git", subdirectory = "models/model_a", tag = "v0.1.0" }
insurance-model-b = { git = "ssh://git@github.com/Mladja99/mock_models.git", subdirectory = "models/model_b", tag = "v0.1.0" }
insurance-model-c = { git = "ssh://git@github.com/Mladja99/mock_models.git", subdirectory = "models/model_c", tag = "v0.1.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
